+++ R/vistime.R
@@ -1,39 +1,73 @@
-#' Time data that is provided is distributed and grouped in a non-overlapping matter. It can then be edited via \code{plotly_build()} and e.g. be used in Shiny apps. The process works offline.
+#' Time data that is provided is distributed and grouped in a non-overlapping matter. It can then be edited via \code{plotly_build()} and used in Shiny apps or Rmarkdown documents. The process works offline.
 #'
 #' @param data data.frame that contains the data to be visualised
-#' @param start (optional) the column name in \code{data} that contains start dates
-#' @param end (optional) the column name in \code{data} that contains end dates
-#' @param groups (optional) the column name in \code{data} to be used for grouping
-#' @param events (optional) the column name in \code{data} that contains event names
-#' @param colors (optional) the column name in \code{data} that contains colors for events
+#' @param events (optional) the column name in \code{data} that contains event names. Default: \code{event}.
+#' @param start (optional) the column name in \code{data} that contains start dates. Default: \code{start}.
+#' @param end (optional) the column name in \code{data} that contains end dates. Default: \code{end}.
+#' @param groups (optional) the column name in \code{data} to be used for grouping. Default: \code{group}.
+#' @param colors (optional) the column name in \code{data} that contains colors for events. Default: \code{color}, if not present, colors are chosen via \code{RColorBrewer}.
+#' @param fontcolors (optional) the column name in \code{data} that contains the font color for event labels. Default: \code{fontcolor}, if not present, color will be \code{black}.
+#' @param tooltips (optional) the column name in \code{data} that contains the mouseover tooltips for the events. Default: \code{tooltip}, if not present, then tooltips are build from event name and date.
+#' @param title (optional) the title to be shown on top of the timeline
 #' @import plotly
 #' @export vistime
 #' @return \code{vistime} returns an object of class "\code{plotly}" and "\code{htmlwidget}".
 #' @examples
-#' # basic example
-#' data(school)
-#' vistime(school, events = "Language", groups = "Room")
-#'
-#' # choose your own colors
+#' # presidents and vice presidents
 #' dat <- data.frame(Position=c(rep("President", 3), rep("Vice", 3)),
 #'                   Name = c("Washington", "Adams", "Jefferson", "Adams", "Jefferson", "Burr"),
 #'                   start = rep(c("1789-03-29", "1797-02-03", "1801-02-03"), 2),
 #'                   end = rep(c("1797-02-03", "1801-02-03", "1809-02-03"), 2),
-#'                   colors = c('#cbb69d', '#603913', '#c69c6e'))
+#'                   color = c('#cbb69d', '#603913', '#c69c6e'),
+#'                   fontcolor = rep("white", 6))
 #'
-#' vistime(dat, events="Position", groups="Name")
-vistime <- function(data, start="start", end="end", groups="group", events="event", colors=NULL){
+#' vistime(dat, events="Position", groups="Name", title="Presidents of the USA")
+#'
+#' # more complex and colorful example
+#' data <- read.csv(text="event,group,start,end,color
+#' Phase 1,Project,2016-12-22,2016-12-23,#c8e6c9
+#' Phase 2,Project,2016-12-23,2016-12-29,#a5d6a7
+#' Phase 3,Project,2016-12-29,2017-01-06,#fb8c00
+#' Phase 4,Project,2017-01-06,2017-02-02,#DD4B39
+#' Start,Start/today,2016-12-22,2016-12-23,#000000
+#' today (after 32 days),Start/today,2017-01-23,2017-01-24,#DD4B39
+#' 1-217.0,category 2,2016-12-27,2016-12-27,#90caf9
+#' 3-200,category 1,2016-12-25,2016-12-25,#1565c0
+#' 3-330,category 1,2016-12-25,2016-12-25,#1565c0
+#' 3-223,category 1,2016-12-28,2016-12-28,#1565c0
+#' 3-225,category 1,2016-12-28,2016-12-28,#1565c0
+#' 3-226,category 1,2016-12-28,2016-12-28,#1565c0
+#' 3-226,category 1,2017-01-19,2017-01-19,#1565c0
+#' 3-330,category 1,2017-01-19,2017-01-19,#1565c0
+#' 3-399.7,moon rising,2017-01-13,2017-01-13,#f44336
+#' 8-831.0,sundowner drink,2017-01-17,2017-01-17,#8d6e63
+#' 9-984.1,birthday party,2016-12-22,2016-12-22,#90a4ae
+#' F01.9,Meetings,2016-12-26,2016-12-26,#e8a735
+#' E43,Meetings,2017-01-12,2017-01-12,#e8a735
+#' R63.3,Meetings,2017-01-12,2017-01-12,#e8a735
+#' Z71,Meetings,2017-01-12,2017-01-12,#e8a735
+#' B95.7,Meetings,2017-01-15,2017-01-15,#e8a735
+#' T82.7,Meetings,2017-01-15,2017-01-15,#e8a735
+#' Room 334,Team 1,2016-12-22,2016-12-28,#DEEBF7
+#' Room 335,Team 1,2016-12-28,2017-01-05,#C6DBEF
+#' Room 335,Team 1,2017-01-05,2017-01-23,#9ECAE1
+#' Group 1,Team 2,2016-12-22,2016-12-28,#E5F5E0
+#' Group 2,Team 2,2016-12-28,2017-01-23,#C7E9C0")
+#'
+#' vistime(data)
+vistime <- function(data, events="event", start="start", end="end", groups="group", colors="color", fontcolors="fontcolor", tooltips="tooltip", title=NULL){
 
-  data <- data.frame(data)
+  data <- data.frame(data, stringsAsFactors = F)
 
   # error checking
-  if(!is.data.frame(data)) stop(paste("Expected an input data frame, but encountered a", class(data)))
+  if(!is.data.frame(data)) stop(paste("Expected an input data frame, but encountered a", class(data)[1]))
+  if(! start %in% names(data)) stop("Please provide the name of the start date column in parameter 'start'")
   if(sum(!is.na(data[, start])) < 1) stop(paste("error in start column: Please provide at least one point in time"))
   if(class(try(as.POSIXct(data[, start]), silent=T))[1] == "try-error") stop(paste("date format error: please provide full dates"))
   if(! events %in% names(data)) stop("Please provide the name of the events column in parameter 'events'")
-  if(! start %in% names(data)) stop("Please provide the name of the start date column in parameter 'start'")
-  if(! groups %in% names(data)) data$group <- ""
+  if(! groups %in% names(data)) data$group <- "" else if(any(is.na(data[, groups]))) stop("if using groups argument, all groups must be set to a non-NA value")
   if(! end %in% names(data)) data$end <- data[, start]
+  if(! (is.null(title) || class(title) %in% c("character", "numeric"))) stop("Title must be a String")
 
   # set column names
   if(events == groups){
@@ -44,9 +78,10 @@ vistime <- function(data, start="start", end="end", groups="group", events="even
   names(data)[names(data)==start] <- "start"
   names(data)[names(data)==end] <- "end"
   names(data)[names(data)==events] <- "event"
+  names(data)[names(data)==events] <- "event"
 
   # sort out the classes
-  data <- as.data.frame(sapply(data, as.character), stringsAsFactors=F)
+  if(nrow(data) > 1){ data <- as.data.frame(sapply(data, as.character), stringsAsFactors=F) }
   data$start <- as.POSIXct(data$start)
   data$end <- as.POSIXct(data$end)
 
@@ -54,62 +89,71 @@ vistime <- function(data, start="start", end="end", groups="group", events="even
   if(any(is.na(data$end))) data$end[is.na(data$end)] <- data$start[is.na(data$end)]
 
   # set the tooltips
-  data$tooltip <- ifelse(data$start == data$end,
-                         paste0("<b>",data$event,": ",data$start,"</b>"),
-                         paste0("<b>",data$event,":</b> from <b>",data$start,"</b> to <b>",data$end,"</b>"))
+  if(tooltips %in% names(data)){
+    names(data)[names(data) == tooltips] <- "tooltip"
+  }else{
+    data$tooltip <- ifelse(data$start == data$end,
+                           paste0("<b>",data$event,": ",data$start,"</b>"),
+                           paste0("<b>",data$event,":</b> from <b>",data$start,"</b> to <b>",data$end,"</b>"))
+  }
 
   # set the colors
-  if(is.null(colors)){
-    palette <- "Set3"
-    data$col <- rep(RColorBrewer::brewer.pal(min(12, max(3, nrow(data))), palette), nrow(data))[1:nrow(data)]
-  }else{
+  if(colors %in% names(data)){
     names(data)[names(data)==colors] <- "col"
+  }else{
+    palette <- "Set3"
+    data$col <- rep(RColorBrewer::brewer.pal(min(11, max(3, nrow(data))), palette), nrow(data))[1:nrow(data)]
+  }
+  if(fontcolors %in% names(data)){
+    names(data)[names(data)==fontcolors] <- "fontcol"
+  }else{
+    data$fontcol <- "black"
   }
 
   ########################################################################
-  #  1. Determine the correct subplot for each event                ######
+  #  1. Determine the correct subplots                              ######
+  #  subplot = groups, but events and ranges separate
   ########################################################################
 
-  data$subplot <- as.numeric(as.factor(data$group))
+  events <- data$start == data$end
+  ranges <- !events
 
+  data$subplot[events] <- as.numeric(factor(data$group[events], levels=unique(data$group[events])))
+  data$subplot[!events] <- length(unique(data$subplot[events])) + as.numeric(factor(data$group[!events], levels=unique(data$group[!events])))
 
   ########################################################################
   #  2. set y values                                                ######
   ########################################################################
-  data <- data[with(data, order(group, start)),]
+  data <- data[with(data, order(subplot, start)),] # order by "start"
   row.names(data) <- 1:nrow(data)
 
   for(sp in unique(data$subplot)){
-    next.y <- 1
 
     # subset data for this group
     thisData <- subset(data, subplot == sp)
-    thisData$y <- 1:nrow(thisData) # base case
+    thisData$y <- 0
 
+    # for each event and for each y, check if any range already drawn on y cuts this range -> if yes, check next y
     for(row in (1:nrow(thisData))){
-      toAdd <- thisData[row,]
+      toAdd <- thisData[row, c("start", "end", "y")]
 
-      # for events: set on new level if it occurs on the same day as previous
-      if(toAdd$start == toAdd$end){
-        if(row>1 && toAdd$start == thisData[row-1, "start"]){
-          next.y <- next.y + 1
+      for(y in 1:nrow(thisData)){
+        thisData[row, "y"] <- y # naive guess
+        # Events
+        if(toAdd$start == toAdd$end){
+          # set on new level if this y is occupied
+          if(all(toAdd$start != thisData[-row,"start"][thisData[-row,"y"] == y])) break; # this y is free, end of search
         }else{
-          next.y <- 1
+          # Ranges, use that already sorted
+          if(all(toAdd$start >= thisData[-row,"end"][thisData[-row,"y"] == y])) break; # new start >= all other starts on this level, end search
         }
-        # for ranges: if this event starts before previous ends, set on new y level (up or below)
-      }else if(row>1 && toAdd$start < thisData[row-1, "end"]){
-        if(next.y == 2){
-          if(toAdd$start >= thisData[thisData$y==1, "end"]){
-            next.y <- next.y-1
-          }else next.y <- next.y + 1
-        }else next.y <- next.y + 1
-      }else{
-        next.y <- 1
       }
-      data[data$subplot == sp, "y"][row] <- next.y
     }
+    data[data$subplot == sp, "y"] <- thisData$y
   }
 
+  data$y <- as.numeric(data$y) # to ensure plotting goes smoothly
+  data$y[is.na(data$y)] <- max(data$y[!is.na(data$y)]) + 1 # just in case
 
   ###########################################################################
   #  3. Set "intelligent" labels for events                           #######
@@ -133,6 +177,8 @@ vistime <- function(data, start="start", end="end", groups="group", events="even
     interval <- 60*10 # 10-min-intervals
   }else if(total_range < 60*60*24){ # max 1 day
     interval <- 60*60*2 # 2-hour-intervals
+  }else if(total_range < 60*60*24*365/2){ # max 0.5 years
+    interval <- 60*60*24 # 1-day-intervals
   }else if(total_range < 60*60*24*365){ # max 1 year
     interval <- 60*60*24*7 # 1-week-intervals
   }else if(total_range < 60*60*24*365*10){ # max 20 years
@@ -153,7 +199,7 @@ vistime <- function(data, start="start", end="end", groups="group", events="even
     thisData <- subset(data, start != end & subplot == sp)
     maxY <- max(thisData$y) + 1
 
-    p <- plot_ly(type = "scatter", mode="lines")
+    p <- plot_ly(data, type = "scatter", mode="lines")
 
     # 1. add vertical line for each year/day
     for(day in seq(min(data$start), max(data$end), interval)){
@@ -166,7 +212,6 @@ vistime <- function(data, start="start", end="end", groups="group", events="even
       toAdd <- thisData[i,]
 
       p <- add_trace(p,
-                     mode="lines",
                      x = c(toAdd$start, toAdd$end),  # von, bis
                      y = toAdd$y,
                      line = list(color = toAdd$col, width = 20),
@@ -175,7 +220,7 @@ vistime <- function(data, start="start", end="end", groups="group", events="even
                      text=toAdd$tooltip) %>%
         add_text(x = toAdd$start + (toAdd$end-toAdd$start)/2,  # in der Mitte
                  y = toAdd$y,
-                 textfont = list(family = "sans serif", size = 14, color = toRGB("black")),
+                 textfont = list(family = "Arial", size = 14, color = toRGB(toAdd$fontcol)),
                  textposition = "center",
                  showlegend=F,
                  text=toAdd$label,
@@ -183,7 +228,7 @@ vistime <- function(data, start="start", end="end", groups="group", events="even
     }
 
     return(p %>% layout(hovermode = 'closest',
-                        margin = list(l=max(nchar(data$group)) * 10),
+                        margin = list(l=max(nchar(data$group)) * 7),
                         # Axis options:
                         # 1. Remove gridlines
                         # 2. Customize y-axis tick labels and show group names instead of numbers
@@ -203,6 +248,7 @@ vistime <- function(data, start="start", end="end", groups="group", events="even
   #######################################################################
 
   eventNumbers <- unique(subset(data, start == end)$subplot)
+
   events <- lapply(eventNumbers, function(sp) {
     # subset data for this Category
     thisData <- subset(data, start == end & subplot == sp)
@@ -225,12 +271,12 @@ vistime <- function(data, start="start", end="end", groups="group", events="even
                      showlegend = F, hoverinfo="text", text=~tooltip)
 
     # add annotations
-    p <- add_text(p, x=~start, y=~y, textfont = list(family = "sans serif", size = 14, color = toRGB("black")),
+    p <- add_text(p, x=~start, y=~y, textfont = list(family = "Arial", size = 14, color = toRGB(toAdd$fontcol)),
                   textposition = ~labelPos, showlegend=F, text = ~label, hoverinfo="none")
 
     # fix layout
     p <-  layout(p, hovermode = 'closest',
-                 margin = list(l=max(nchar(data$group)) * 10),
+                 margin = list(l=max(nchar(data$group)) * 7),
                  xaxis = list(showgrid = F, title=''),
                  yaxis = list(showgrid = F, title = '',
                               tickmode = "array", tickvals = 1:maxY,
@@ -253,7 +299,7 @@ vistime <- function(data, start="start", end="end", groups="group", events="even
   # gather all plots in a plotList
   plotList <- append(ranges, events)
 
-  total <- subplot(plotList, nrows=length(plotList), shareX=T, margin=0, heights=heightsRelative)
+  total <- subplot(plotList, nrows=length(plotList), shareX=T, margin=0, heights=heightsRelative) %>% layout(title = title)
 
   return(total)